mule player count

+++ common/database.cpp
@@ -452,10 +452,10 @@ bool Database::SetMule(const char* charname) {
 		return false;
 	}
 	auto row = results.begin();
-	uint8 account_id = std::stoi(row[0]);
+	uint32 account_id = std::stoul(row[0]);
 	
 	// iterate over every character associated with account and verify they're all level 1
-	std::string query = StringFormat("SELECT `account_id`, `level` FROM `character_data` WHERE `account_id` = %d", account_id);
+	std::string query = StringFormat("SELECT `account_id`, `level` FROM `character_data` WHERE `account_id` = %u", account_id);
 	auto results = QueryDatabase(query);
 	for (auto row = results.begin(); row != results.end(); ++row) {
 		if (std::stoi(row[1]) != 1) {
@@ -465,7 +465,7 @@ bool Database::SetMule(const char* charname) {
 	}
 
 	// finally set account to mule status
-	std::string query = StringFormat("UPDATE account SET mule = %d where id = %d AND status < 80", 1, account_id);
+	std::string query = StringFormat("UPDATE account SET mule = %d where id = %u AND status < 80", 1, account_id);
 	auto results = QueryDatabase(query);
 	if (!results.Success()) {
 		return false;
